Remove commits made for VS2019easy.
In making the VS2019easy branch easy to
compile on windows, I embedded a bunch
of commits to make compilation straightforward
without having to use an external CMake, for
ease of compilation for the students of my class.

These should not be in the mainline of development,
which I had made them be. Fix that.

Revert "embeddeding resources into exe."

This reverts commit ba1861377b1417c7188a5ac52f8589a74fa555ae.