handle the gui on/off more cleanly.
Before, imgui was being called to set up the rendering,
just not to do it if escape had been pressed.

This techinque was implemented because I was occasionally
getting IMGUI errors, I assume because their data structures
could be left in an unfinished state because of pressing escape,
or resuming.